Unreviewed build fix for 32-bit builds.
Make casts from unsigned long long -> size_t explicit, rather than implicit. Also
make CARingBuffer be protected by USING_MEDIA_TOOLBOX so ports without that flag
won't try to compile unused code.

* platform/audio/mac/CARingBuffer.cpp:
(WebCore::CARingBuffer::fetch):
* platform/audio/mac/CARingBuffer.h:
* platform/graphics/avfoundation/AudioSourceProviderAVFObjC.mm:
(WebCore::AudioSourceProviderAVFObjC::provideInput):

Canonical link: https://commits.webkit.org/154660@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@173634 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/platform/graphics/avfoundation/AudioSourceProviderAVFObjC.mm
@@ -119,7 +119,7 @@ void AudioSourceProviderAVFObjC::provideInput(AudioBus* bus, size_t framesToProc
         return;
     }
 
-    size_t framesAvailable = endFrame - (m_readCount + m_writeAheadCount);
+    size_t framesAvailable = static_cast<size_t>(endFrame - (m_readCount + m_writeAheadCount));
     if (framesAvailable < framesToProcess) {
         framesToProcess = framesAvailable;
         bus->zero();
